+++ Sofa/framework/DefaultType/src/sofa/defaulttype/SolidTypes.h
@@ -31,7 +31,7 @@
 #include <sofa/type/vector.h>
 #include <iostream>
 #include <map>
-#include <sofa/defaulttype/SpatialVector.h>
+#include <sofa/defaulttype/Transform.h>
 
 
 
@@ -43,10 +43,6 @@ Base types for the ArticulatedSolid: position, orientation, velocity, angular ve
 
 @author François Faure, INRIA-UJF, 2006
 */
-
-class SOFA_DEFAULTTYPE_API Transform;
-
-
 template< class R=float >
 class SOFA_DEFAULTTYPE_API SolidTypes
 {
@@ -63,6 +59,7 @@ public:
     typedef Vec6 DOF; ///< For compatibility
 
     using SpatialVector = SpatialVector<R>;
+    using Transform = Transform<R>;
 
     /**
      * \brief A twist aka a SpatialVector representing a velocity
@@ -88,144 +85,6 @@ public:
     public:
         Wrench(const Vec3& force, const Vec3& torque)
             : SpatialVector(force, torque) {}
-    };
-
-    /** Define a frame (child) whith respect to another (parent). A frame represents a local coordinate system.
-
-    Internal data represents the orientation of the child wrt the parent, BUT the translation vector represents the origin of the parent with respect to the child. For example, the coordinates M_p of point M in parent given the coordinates M_c of the same point in child are given by: M_p = orientation * ( M_c - origin ). This is due to Featherstone's conventions. Use method setTranslationRotation( const Vec& t, const Rot& q ) to model the Transform the standard way (i.e. translation givne in the parent frame).
-
-
-    */
-    class SOFA_DEFAULTTYPE_API Transform
-    {
-    public:
-        /// The default constructor does not initialize the transform
-        Transform();
-        /// Origin of the child in parent coordinates, orientation of the child wrt to parent
-        Transform( const Vec& origin, const Rot& orientation );
-        /// WARNING: using Featherstone's conventions (see class documentation)
-        Transform( const Rot& q, const Vec& o );
-        /// Origin of the child in the parent coordinate system and the orientation of the child wrt the parent (i.e. standard way)
-        void set( const Vec& t, const Rot& q );
-        /// Reset this to identity
-        void clear();
-        /// The identity transform (child = parent)
-        static Transform identity();
-        /// Origin of the child in the parent coordinate system and the orientation of the child wrt the parent (i.e. standard way)
-        //static Transform inParent(const Vec& t, const Rot& r);
-        /// Define child as a given SpatialVector integrated during one second, starting from the parent (used for time integration). The spatial vector is given in parent coordinates.
-        Transform( const SpatialVector& v );
-        /// The inverse transform i.e. parent wrt child
-        Transform inversed() const;
-        /// Parent origin in child coordinates (the way it is actually stored internally)
-        const Vec& getOriginOfParentInChild() const;
-        /// Origin of child in parent coordinates
-        Vec getOrigin() const;
-        /// Origin of child in parent coordinates
-        void setOrigin( const Vec& );
-        /// Orientation of the child coordinate axes wrt the parent coordinate axes
-        const Rot& getOrientation() const;
-        /// Orientation of the child coordinate axes wrt the parent coordinate axes
-        void setOrientation( const Rot& );
-        /// Matrix which projects vectors from child coordinates to parent coordinates. The columns of the matrix are the axes of the child base axes in the parent coordinate system.
-        Mat3x3 getRotationMatrix() const;
-
-
-
-
-
-
-        /**
-         * \brief Adjoint matrix to the transform
-         * This matrix transports velocities in twist coordinates from the child frame to the parent frame.
-         * Its inverse transpose does the same for the wrenches
-         */
-        Mat6x6 getAdjointMatrix() const;
-
-        /// Project a vector (i.e. a direction or a displacement) from child coordinates to parent coordinates
-        Vec projectVector( const Vec& vectorInChild ) const;
-        /// Project a point from child coordinates to parent coordinates
-        Vec projectPoint( const Vec& pointInChild ) const;
-        /// Projected a vector (i.e. a direction or a displacement) from parent coordinates to child coordinates
-        Vec backProjectVector( const Vec& vectorInParent ) const;
-        /// Project point from parent coordinates to this coordinates
-        Vec backProjectPoint( const Vec& pointInParent ) const;
-        /// Combine two transforms. If (*this) locates frame B (child) wrt frame A (parent) and if f2 locates frame C (child) wrt frame B (parent) then the result locates frame C wrt to Frame A.
-        Transform operator * (const Transform& f2) const;
-        /// Combine two transforms. If (*this) locates frame B (child) wrt frame A (parent) and if f2 locates frame C (child) wrt frame B (parent) then the result locates frame C wrt to Frame A.
-        Transform& operator *= (const Transform& f2);
-
-        /** Project a spatial vector from child to parent
-            *  TODO One should handle differently the transformation of a twist and a wrench !
-            *  This applying the adjoint to velocities or its transpose to wrench :
-            *  V_parent = Ad . V_child or W_child = Ad^T . W_parent
-            *  To project a wrench in the child frame to the parent frame you need to do
-            *  parent_wrench = this->inversed * child_wrench
-            *  (this doc needs to be douv-ble checked !)
-            */
-        // create a spatial Vector from a small transformation
-        SpatialVector  CreateSpatialVector();
-        SpatialVector DTrans();
-
-        SpatialVector operator * (const SpatialVector& sv ) const;
-        /// Project a spatial vector from parent to child (the inverse of operator *). This method computes (*this).inversed()*sv without inverting (*this).
-        SpatialVector operator / (const SpatialVector& sv ) const;
-        /// Write an OpenGL matrix encoding the transformation of the coordinate system of the child wrt the coordinate system of the parent.
-        void writeOpenGlMatrix( double *m ) const;
-        /// Draw the axes of the child coordinate system in the parent coordinate system
-        /// Print the origin of the child in the parent coordinate system and the quaternion defining the orientation of the child wrt the parent
-        inline friend std::ostream& operator << (std::ostream& out, const Transform& t )
-        {
-            out << t.getOrigin() << " " << t.getOrientation();
-            return out;
-        }
-
-        /// read from an input stream
-        inline friend std::istream& operator >> ( std::istream& in, Transform& t )
-        {
-            Vec origin;
-            Rot orientation;
-
-            in >> origin >> orientation;
-
-            t.set(origin, orientation);
-
-            return in;
-        }
-
-        /// Print the internal values (i.e. using Featherstone's conventions, see class documentation)
-        void printInternal( std::ostream&) const;
-
-        /** @name Time integration
-        * Methods used in time integration
-        */
-        ///@{
-        /// (*this) *= Transform(v)  Used for time integration. SHOULD WE RATHER APPLY (*this)=Transform(v)*(*this) ???
-        Transform& operator +=(const SpatialVector& a);
-
-        Transform& operator +=(const Transform& a);
-
-        template<class Real2>
-        Transform& operator*=(Real2 a)
-        {
-            origin_ *= a;
-            return *this;
-        }
-
-        template<class Real2>
-        Transform operator*(Real2 a) const
-        {
-            Transform r = *this;
-            r*=a;
-            return r;
-        }
-        ///@}
-
-
-    protected:
-        Rot orientation_; ///< child wrt parent
-        Vec origin_;  ///< parent wrt child
-
     };
 
 
@@ -299,7 +158,6 @@ public:
 #if !defined(SOFA_DEFAULTTYPE_SOLIDTYPES_CPP)
 extern template class SOFA_DEFAULTTYPE_API SolidTypes<double>;
 extern template class SOFA_DEFAULTTYPE_API SolidTypes<float>;
-
 #endif
 
 }
